requirement.txt
Pillow==9.5.0
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pyperclip==1.8.2
pystray==0.19.5

